+++ CHANGELOG.md
@@ -6,7 +6,8 @@ A summary of changes made to the codebase, grouped per deployment.
 
 ### New features
 
-- 
+- Created page showing statistics about the use of the workshop and reservations of the equipment
+
 
 ### Improvements
 
@@ -31,6 +32,7 @@ A summary of changes made to the codebase, grouped per deployment.
     - _Examples can be passwords of Google Workspace users that only exist for technical purposes,
       or credentials for the admin page of a physical router_
 - Added an "internal" ribbon label on internal articles and events
+>>>>>>> 66675ec2 (Added page to show statistics from use of the workshop)
 
 ### Improvements